+++ lib/functions/sparse_connectedpixels/sparse_connectedpixels.c
@@ -7,14 +7,15 @@ static int NOISY = 0;
  * {"py_sig": "sparse_connectedpixels(v: buffer, i: buffer, j: buffer, threshold: float, labels: buffer) -> int",
  *  "doc": "runs the connectedpixels algorithm on\na sparse image using a supplied threshold putting labels\ninto labels array and returning the number of blobs found",
  *  "params": {"v": "Values (float32).", "i": "Rows (uint16).", "j": "Cols (uint16).", "threshold": "Threshold.", "labels": "Output labels (int32)."},
- *  "checks": ["v.format == 'f'", "i.format == 'H' or i.itemsize == 2", "j.format == 'H' or j.itemsize == 2",
+ *  "checks": ["v.slow_axis == 0", "i.format == 'H' or i.itemsize == 2", "j.format == 'H' or j.itemsize == 2",
  *      "j.n == i.n", "v.n == i.n", "( labels.format == 'i' or labels.format == 'l' )", "labels.n == i.n"],
- *  "c_overloads": [{"sig": "int sparse_connectedpixels(const float *v, const uint16_t *i, const uint16_t *j, intptr_t nnz, float threshold, int32_t *labels) -> int",
+ *  "c_overloads": [{"when": "v.format == 'f' and v.slow_axis == 0",
+  *         "sig": "int sparse_connectedpixels(const float *v, const uint16_t *i, const uint16_t *j, intptr_t nnz, double threshold, int32_t *labels) -> int",
  *      "map": {"v": "v.ptr", "i": "i.ptr", "j": "j.ptr", "nnz": "i.n", "threshold": "threshold", "labels": "labels.ptr"}}]}
 C2PY_END */
 
 int sparse_connectedpixels(float *restrict v, uint16_t *restrict i,
-                           uint16_t *restrict j, intptr_t nnz, float threshold,
+                            uint16_t *restrict j, intptr_t nnz, double threshold,
                            int32_t *restrict labels /* nnz */
 ) {
     intptr_t k, p, pp; int ir;
